chore: remove circular import with options and utils
